backend/requirements.txt
# ============================================================================
# CORE WEB FRAMEWORK
# ============================================================================
fastapi==0.110.1
uvicorn[standard]==0.25.0
python-multipart==0.0.9
starlette==0.37.2

# ============================================================================
# BIOINFORMATICS
# ============================================================================
biopython==1.83

# ============================================================================
# DATA PROCESSING & VALIDATION
# ============================================================================
pandas
numpy
pydantic==2.5.0
pydantic-settings==2.1.0

# ============================================================================
# CONFIGURATION
# ============================================================================
python-dotenv==1.0.0

# ============================================================================
# MACHINE LEARNING
# ============================================================================
scikit-learn
joblib

# ============================================================================
# LANGGRAPH & ORCHESTRATION (NEW)
# ============================================================================
langgraph
langchain
langchain-community

# ============================================================================
# LLAMAINDEX & RAG (NEW)
# ============================================================================
llama-index
llama-index-embeddings-openai

# ============================================================================
# GEMINI INTEGRATION
# ============================================================================
google-generativeai
google-api-core

# ============================================================================
# VECTOR DATABASES
# ============================================================================
chroma==0.4.10
chromadb==0.4.10

# ============================================================================
# UTILITIES
# ============================================================================
python-json-logger==2.0.7
colorama==0.4.6
requests==2.31.0
